fix(DataControl): lift fetchFiles to the Home page, update files when a new file is uploaded
